Actually fix underline issue
This seems like a bug in Safari, where if at any point the hover rule and the data-attribute rule
overlapped, it would get "stuck" in a competing hover state where it rendered both colors,
even when the element was no longer being hovered over.

Using :not(:hover) did not work either, for some reason.
The only solution I found was using the selector :hover:not([data-attr])